+++ src/seedsigner/views/seed_views.py
@@ -16,7 +16,7 @@ from seedsigner.gui.screens import (RET_CODE__BACK_BUTTON, ButtonListScreen,
 from seedsigner.gui.screens.screen import LargeIconStatusScreen, QRDisplayScreen
 from seedsigner.helpers import embit_utils
 from seedsigner.models.decode_qr import DecodeQR
-from seedsigner.models.encode_qr import EncodeQR
+from seedsigner.helpers.qr_encoders import CompactSeedQrEncoder, GenericStaticQrEncoder, SeedQrEncoder, SpecterXPubQrEncoder, StaticXpubQrEncoder, UrXpubQrEncoder
 from seedsigner.models.psbt_parser import PSBTParser
 from seedsigner.models.qr_type import QRType
 from seedsigner.models.seed import InvalidSeedException, Seed
@@ -863,34 +863,24 @@ class SeedExportXpubQRDisplayView(View):
         super().__init__()
         self.seed = self.controller.get_seed(seed_num)
 
-        qr_density = self.settings.get_value(SettingsConstants.SETTING__QR_DENSITY)
-        if coordinator == SettingsConstants.COORDINATOR__SPECTER_DESKTOP:
-            qr_type = QRType.XPUB__SPECTER
-
-        elif coordinator == SettingsConstants.COORDINATOR__BLUE_WALLET:
-            qr_type = QRType.XPUB
-
-        elif coordinator == SettingsConstants.COORDINATOR__KEEPER:
-            qr_type = QRType.XPUB
-
-        elif coordinator == SettingsConstants.COORDINATOR__NUNCHUK:
-            qr_type = QRType.XPUB__UR
-
-            # As of 2022-03-02 Nunchuk doesn't seem to support animated QRs for Xpub import
-            qr_density = SettingsConstants.DENSITY__HIGH
-
-        else:
-            qr_type = QRType.XPUB__UR
-
-        self.qr_encoder = EncodeQR(
+        encoder_args = dict(
             seed_phrase=self.seed.mnemonic_list,
             passphrase=self.seed.passphrase,
             derivation=derivation_path,
             network=self.settings.get_value(SettingsConstants.SETTING__NETWORK),
-            qr_type=qr_type,
-            qr_density=qr_density,
-            wordlist_language_code=self.seed.wordlist_language_code
+            qr_density=self.settings.get_value(SettingsConstants.SETTING__QR_DENSITY)
         )
+
+        if coordinator == SettingsConstants.COORDINATOR__SPECTER_DESKTOP:
+            self.qr_encoder = SpecterXPubQrEncoder(**encoder_args)
+
+        elif coordinator in [SettingsConstants.COORDINATOR__BLUE_WALLET,
+                             SettingsConstants.COORDINATOR__KEEPER,
+                             SettingsConstants.COORDINATOR__NUNCHUK]:
+            self.qr_encoder = StaticXpubQrEncoder(**encoder_args)
+
+        else:
+            self.qr_encoder = UrXpubQrEncoder(**encoder_args)
 
 
     def run(self):
@@ -1391,11 +1381,13 @@ class SeedTranscribeSeedQRWholeQRView(View):
     
 
     def run(self):
-        e = EncodeQR(
-            seed_phrase=self.seed.mnemonic_list,
-            qr_type=self.seedqr_format,
-            wordlist_language_code=self.settings.get_value(SettingsConstants.SETTING__WORDLIST_LANGUAGE)
-        )
+        encoder_args = dict(mnemonic=self.seed.mnemonic_list,
+                            wordlist_language_code=self.settings.get_value(SettingsConstants.SETTING__WORDLIST_LANGUAGE))
+        if self.seedqr_format == QRType.SEED__SEEDQR:
+            e = SeedQrEncoder(**encoder_args)
+        elif self.seedqr_format == QRType.SEED__COMPACTSEEDQR:
+            e = CompactSeedQrEncoder(**encoder_args)
+
         data = e.next_part()
 
         ret = seed_screens.SeedTranscribeSeedQRWholeQRScreen(
@@ -1426,11 +1418,13 @@ class SeedTranscribeSeedQRZoomedInView(View):
     
 
     def run(self):
-        e = EncodeQR(
-            seed_phrase=self.seed.mnemonic_list,
-            qr_type=self.seedqr_format,
-            wordlist_language_code=self.settings.get_value(SettingsConstants.SETTING__WORDLIST_LANGUAGE)
-        )
+        encoder_args = dict(seed_phrase=self.seed.mnemonic_list,
+                            wordlist_language_code=self.settings.get_value(SettingsConstants.SETTING__WORDLIST_LANGUAGE))
+        if self.seedqr_format == QRType.SEED__SEEDQR:
+            e = SeedQrEncoder(**encoder_args)
+        elif self.seedqr_format == QRType.SEED__COMPACTSEEDQR:
+            e = CompactSeedQrEncoder(**encoder_args)
+
         data = e.next_part()
 
         if len(self.seed.mnemonic_list) == 24:
@@ -2071,7 +2065,7 @@ class SeedSignMessageSignedMessageQRView(View):
 
 
     def run(self):
-        qr_encoder = EncodeQR(qr_type=QRType.SIGN_MESSAGE, signed_message=self.signed_message)
+        qr_encoder = GenericStaticQrEncoder(data=self.signed_message)
         
         self.run_screen(
             QRDisplayScreen,
