+++ tests/test_flows_psbt.py
@@ -1,6 +1,5 @@
 from base import FlowTest, FlowStep
 
-from seedsigner.controller import Controller
 from seedsigner.views.view import MainMenuView
 from seedsigner.views import scan_views, seed_views, psbt_views
